requirements.txt
requests
python-dotenv
SQLAlchemy>=2
pandas
tenacity
loguru
